Move the dashboard-only components into src/components
DateRangePicker, PageHeader, StatTile and FormPasswordInput have no equivalent
in the app, so they move out of the page folder rather than being replaced.
Nothing about them is dashboard-specific.

toIsoDate and isoDaysAgo move to lib/utils with them: DateRangePicker's presets
need them, and a component under src/components has no business importing from
a page.
