+++ frontend/src/components/FormPasswordInput/index.tsx
@@ -1,7 +1,7 @@
 import { useId, useState } from "react";
 import { Eye, EyeOff } from "lucide-react";
 import { Input } from "@components/Shadcn/Input";
-import { cn } from "@pages/business-dashboard/lib/utils";
+import { cn } from "@/lib/utils";
 import type { IProps } from "./types";
 
 const FormPasswordInput = ({ label, registration, error, hint, className, ...props }: IProps) => {
