+++ frontend/src/components/StatTile/types.ts
