+++ frontend/src/components/StatTile/variants.ts
