+++ frontend/src/pages/business-dashboard/pages/overview/KpiRow.tsx
@@ -1,5 +1,5 @@
 import { CircleCheck, FileText, Layers, XCircle } from "lucide-react";
-import StatTile from "@pages/business-dashboard/components/StatTile";
+import StatTile from "@components/StatTile";
 import { formatCompact, formatPercent } from "@pages/business-dashboard/lib/utils";
 import type { SessionStatsResponse } from "@pages/business-dashboard/services/types";
