+++ frontend/src/pages/business-dashboard/pages/sessions/FilterBar.tsx
@@ -1,6 +1,6 @@
 import { RotateCcw, Search } from "lucide-react";
 import { Button } from "@components/Shadcn/Button";
-import DateRangePicker from "@pages/business-dashboard/components/DateRangePicker";
+import DateRangePicker from "@components/DateRangePicker";
 import { Input } from "@components/Shadcn/Input";
 import Select, {
     SelectContent,
