+++ frontend/src/pages/business-dashboard/pages/sessions/index.tsx
@@ -3,7 +3,7 @@ import { ROUTES } from "@constants/routes";
 import { Download, RefreshCw } from "lucide-react";
 
 import { Button } from "@components/Shadcn/Button";
-import PageHeader from "@pages/business-dashboard/components/PageHeader";
+import PageHeader from "@components/PageHeader";
 import { useSessionsPage } from "./useSessionsPage";
 import FilterBar from "./FilterBar";
 import PaginationBar from "./PaginationBar";
